Fix file org. inconsistency in `types`
