+++ eth2/types/src/lib.rs
@@ -8,6 +8,7 @@ pub mod beacon_block;
 pub mod beacon_block_body;
 pub mod beacon_state;
 pub mod casper_slashing;
+pub mod chain_spec;
 pub mod crosslink;
 pub mod deposit;
 pub mod deposit_data;
@@ -28,7 +29,6 @@ pub mod slashable_vote_data;
 pub mod slot_epoch_macros;
 pub mod slot_epoch;
 pub mod slot_height;
-pub mod spec;
 pub mod validator;
 pub mod validator_registry;
 pub mod validator_registry_delta_block;
@@ -44,6 +44,7 @@ pub use crate::beacon_block::BeaconBlock;
 pub use crate::beacon_block_body::BeaconBlockBody;
 pub use crate::beacon_state::BeaconState;
 pub use crate::casper_slashing::CasperSlashing;
+pub use crate::chain_spec::ChainSpec;
 pub use crate::crosslink::Crosslink;
 pub use crate::deposit::Deposit;
 pub use crate::deposit_data::DepositData;
@@ -60,7 +61,6 @@ pub use crate::slashable_attestation::SlashableAttestation;
 pub use crate::slashable_vote_data::SlashableVoteData;
 pub use crate::slot_epoch::{Epoch, Slot};
 pub use crate::slot_height::SlotHeight;
-pub use crate::spec::ChainSpec;
 pub use crate::validator::{StatusFlags as ValidatorStatusFlags, Validator};
 pub use crate::validator_registry_delta_block::ValidatorRegistryDeltaBlock;
